% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_bivar.R
\name{entropy_bivar}
\alias{entropy_bivar}
\title{Bivariate Entropies}
\usage{
entropy_bivar(dat)
}
\arguments{
\item{dat}{Dataframe with rows as observations and columns as variables. Variables must all be observed or transformed categorical with finite range spaces.}
}
\value{
Upper triangular matrix giving bivariate entropies between pairs of variables which are given as rows and columns of the matrix.
The univariate entropies are given in the diagonal.
}
\description{
Calculates the bivariate entropies
}
\details{
To be completed
}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data. *Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique*, 129(1), 45-63.
}
\author{
Termeh Shafie
}
