% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dyad_var.R
\name{get_dyad_var}
\alias{get_dyad_var}
\title{Get Dyad Variables}
\usage{
get_dyad_var(var, type = "att")
}
\arguments{
\item{var}{variable vector (actor attribute) or adjacency matrix (ties)
to be transformed to a dyad variable.}

\item{type}{either 'att' for actor attribute (default) or 'tie' for relations.}
}
\value{
Data frame with three columns:
first two columns show the vertex pairs \code{u} and \code{v} where \code{u<v} ,
and the third column gives the transformed dyadic variable \code{var}.
}
\description{
Transforms vertex variables or observed directed/undirected ties
into dyad variables.
}
\details{
For actor attributes, unique pairs of original attribute values
constitute the outcome space. Note that the actor  attributes need
to be categorical and on ordinal scale with finite range spaces.
\cr

For directed relations, pairs of indicators from the adjacency matrix constitute
the four outcomes representing possible combinations of sending and receiving ties:
(0,0), (0,1), (1,0), (1,1).
\cr

For undirected relations, an indicator variable which is directly read from the
adjacency matrix represents the dyadic variable.
\cr

Warning message is shown if actor attribute has too many unique outcomes
as it will yield too many possible outcomes when converted in to a dyad variable.
\cr

To be completed.
}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data.
\emph{Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique}, 129(1), 45-63.
\cr

Nowicki, K., Shafie, T., & Frank, O. (Forthcoming 2022). \emph{Statistical Entropy Analysis of Network Data}.
}
\seealso{
\code{\link{joint_entropy}}
}
\author{
Termeh Shafie
}
