% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_triad_var.R
\name{get_triad_var}
\alias{get_triad_var}
\title{Get Triad Variables}
\usage{
get_triad_var(var, type = "att")
}
\arguments{
\item{var}{variable vector (actor attribute) or adjacency matrix (ties)
to be transformed to a triad variable.}

\item{type}{either 'att' for actor attribute (default) or 'tie' for relations.}
}
\value{
Data frame with four columns:
first three columns show the vertex triad \code{u}, \code{v}, \code{w} ,
and the fourth column gives the transformed triadicvariable \code{var}.
}
\description{
Transforms vertex variables or observed directed/undirected ties
into triad variables.
}
\details{
For actor attributes, unique triples of original attribute values
constitute the outcome space. Note that the actor
attributes need to be categorical and on ordinal scale with finite range spaces.

For directed relations, a sequence of indicators of length 6 created from the adjacency matrix
constitutes the 64 outcomes representing possible combinations of sending and receiving ties.

For undirected relations, triples of indicators are created from the adjacency matrix.

Warning message is shown if actor attribute has too many unique outcomes
as it will yield too many possible outcomes when converted to a triad variable.

To be completed.
}
\examples{
To be added
}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data. *Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique*, 129(1), 45-63.
}
\seealso{

}
\author{
Termeh Shafie
}
