% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_graph.R
\name{assoc_graph}
\alias{assoc_graph}
\title{Association graphs}
\usage{
assoc_graph(dat, cutoff = 0)
}
\arguments{
\item{dat}{Dataframe with rows as observations and columns as variables. Variables must all be observed or transformed categorical with finite range spaces.}

\item{cutoff}{The cutoff point for the edges to be drawn based on joint entropies. Default is 0 and draws all edges.}
}
\value{
A ggraph object.
}
\description{
Draws association graphs based on given thresholds of joint entropy values between pairs of variables.
}
\details{
To be completed
}
\examples{
# to be added
}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data. *Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique*, 129(1), 45-63.
}
\author{
Termeh Shafie
}
