% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_gof.R
\name{div_gof}
\alias{div_gof}
\title{Divergence Tests of Goodness of Fit}
\usage{
div_gof(dat, set1, set2, condition = 0)
}
\arguments{
\item{dat}{dataframe with rows as observations and columns as variables.
Variables must all be observed or transformed categorical with finite range spaces.}

\item{set1}{column vector of variables in the set
to be tested for independence against those in \code{set2}}

\item{set2}{column vector of variables in set
to be tested for independence against those in \code{set1}}

\item{condition}{column vector of variables to condition the independence test on,
must be different variables than those in \code{set1} and set \code{set2}).
Default empty (no conditioning).}

\item{spec}{either 'pairwise', 'conditional' or 'nested' to specify the model to test}
}
\value{
Test results
}
\description{
Tests of various hypothetical structural models \emph{p0} against
the general model \emph{p} which is estimated using empirical data.
}
\details{
description of tests
}
\examples{

}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data.
\emph{Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique}, 129(1), 45-63.
\cr

Nowicki, K., Shafie, T., & Frank, O. (Forthcoming 2022). \emph{Statistical Entropy Analysis of Network Data}.
}
\seealso{
\code{\link{get_dyad_var}}
}
\author{
Termeh Shafie
}
