% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_gof.R
\name{div_gof}
\alias{div_gof}
\title{Divergence Tests of Goodness of Fit}
\usage{
div_gof(dat, var1, var2, var_cond = NULL)
}
\arguments{
\item{dat}{dataframe with rows as observations and columns as variables.
Variables must all be observed or transformed categorical with finite range spaces.}

\item{var1}{character of variable name in \code{dat} specified to be independent to
those specified in \code{var2} under \emph{p0}}

\item{var2}{character of variable name in \code{dat} specified to be independent to
those specified in \code{var1}}

\item{var_cond}{character of variable name in \code{dat} to condition the independence specification on,
must be different variables than those specified in \code{var1} and \code{var2}).
Default = empty (no conditioning).}
}
\value{
Message indicating whether the hypothesis with the specified independence model
can be rejected or not on approximately 5\% level of significance.
\item{summary}{Dataframe including the value of the divergence \emph{D} and its degrees of freedom}
}
\description{
Tests of various hypothetical structural models \emph{p0} against
the general model \emph{p} which is estimated using empirical data.
}
\details{
this function is currently implemented to only test goodness of fit of models specified as
\emph{X} independent of \emph{Y}, or \emph{X} independent of \emph{Y} given \emph{Z}.
}
\examples{

}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data.
\emph{Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique}, 129(1), 45-63.
\cr

Nowicki, K., Shafie, T., & Frank, O. (Forthcoming 2022). \emph{Statistical Entropy Analysis of Network Data}.
}
\seealso{
\code{\link{joint_entropy}},   \code{\link{assoc_graph}},  \code{\link{entropy_trivar}}
}
\author{
Termeh Shafie
}
