% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_gof.R
\name{div_gof}
\alias{div_gof}
\title{Divergence Tests of Goodness of Fit}
\usage{
div_gof(dat, var1, var2, var_cond = 0)
}
\arguments{
\item{dat}{dataframe with rows as observations and columns as variables.
Variables must all be observed or transformed categorical with finite range spaces.}

\item{var1}{vector of variables in \code{dat} specified to be independent to
those specified in \code{var2} under \emph{p0}}

\item{var2}{vector of variables in \code{dat} specified to be independent to
those specified in \code{var1}}

\item{var_cond}{vector of variables in \code{dat} to condition the independence specification on,
must be different variables than those specified in \code{var1} and \code{var2}).
Default empty (no conditioning).}
}
\value{
Test results
}
\description{
Tests of various hypothetical structural models \emph{p0} against
the general model \emph{p} which is estimated using empirical data.
}
\details{
description of tests
}
\examples{

}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data.
\emph{Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique}, 129(1), 45-63.
\cr

Nowicki, K., Shafie, T., & Frank, O. (Forthcoming 2022). \emph{Statistical Entropy Analysis of Network Data}.
}
\seealso{
\code{\link{get_dyad_var}}
}
\author{
Termeh Shafie
}
