% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_gof.R
\name{div_gof}
\alias{div_gof}
\title{Divergence Tests of Goodness of Fit}
\usage{
div_gof(varset1, varset2, condition = 0)
}
\arguments{
\item{varset1}{column vector of variables in set 1}

\item{varset2}{column vector of variables in set 2}

\item{condition}{column vector of variables, default empty (no conditioning)}
}
\value{
Test results
}
\description{
Tests of various hypothetical structural models \emph{p0} against the general the model \emph{p} which is
estimated using the empirical data.
}
\details{
description of tests
}
\examples{

}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data.
\emph{Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique}, 129(1), 45-63.
\cr

Nowicki, K., Shafie, T., & Frank, O. (Forthcoming 2022). \emph{Statistical Entropy Analysis of Network Data}.
}
\seealso{
\code{\link{get_dyad_var}}
}
\author{
Termeh Shafie
}
