% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_trivar.R
\name{entropy_trivar}
\alias{entropy_trivar}
\title{Trivariate Entropy}
\usage{
entropy_trivar(dat)
}
\arguments{
\item{dat}{Dataframe with rows as observations and columns as variables.
Variables must all be observed or transformed categorical with finite range spaces.}
}
\value{
Dataframe containing all possible triples of variables and their entropies.
}
\description{
Computes the trivariate entropies of all triples of (discrete) variables in a multivariate data set.
}
\details{
To be completed
}
\examples{
To be added
}
\references{
Frank, O., & Shafie, T. (2016). Multivariate entropy analysis of network data. *Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique*, 129(1), 45-63.
}
\seealso{
\code{\link{joint_entrorpy}}
}
\author{
Termeh Shafie
}
